html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-block-formatting-context.html

<!DOCTYPE html>
<title>The fieldset element: block formatting context</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
* {
  margin: 0;
  padding: 0;
}
fieldset { border: none; }
.fieldset { -webkit-appearance: fieldset; }
.float {
  float: left;
  width: 50px;
  height: 50px;
  background-color: orange;
}
</style>

<div class=float></div>
<fieldset><div class=float></div></fieldset>

<div class=float></div>
<div class=fieldset><div class=float></div></div>

<script>
let expectedTop = 0;
for (const fieldset of [document.querySelector('fieldset'), document.querySelector('.fieldset')]) {
  test(() => {
    assert_equals(fieldset.offsetTop, expectedTop, 'fieldset.offsetTop');
    assert_equals(fieldset.offsetLeft, 50, 'fieldset.offsetLeft');
    assert_equals(fieldset.clientHeight, 50, 'fieldset.clientHeight');
  });
  expectedTop += 50;
}
</script>
